added new types
