+++ src/types/parent.ts
@@ -1,0 +1,11 @@
+import { Address } from "./address";
+import { User } from "./user";
+
+export interface Parent extends User {
+    parentId: number;
+    fistName: string;
+    lastName: string;
+    gender: string;
+    age: number;
+    address: Address;
+  }
